% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileClass.R
\name{as.character,FileClass-method}
\alias{as.character,FileClass-method}
\title{Convert a FileClass object to a google URI}
\usage{
\S4method{as.character}{FileClass}(x, ...)
}
\arguments{
\item{x}{FileClass object}

\item{...}{not used}
}
\value{
google URI
}
\description{
Convert a FileClass object to a google URI
}
