% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderClass.R
\name{as.character,FolderClass-method}
\alias{as.character,FolderClass-method}
\title{Convert A FolderClass object to a dummy google URI}
\usage{
\S4method{as.character}{FolderClass}(x, ...)
}
\arguments{
\item{x}{FolderClass object}

\item{...}{not used}
}
\value{
google URI
}
\description{
Convert A FolderClass object to a dummy google URI, this URI
does not exist on google, it can only be used in `gcs_cp`.
}
