% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcs_cloud_auth}
\alias{gcs_cloud_auth}
\alias{gcs_get_cloud_auth}
\title{Get/Set google credentials}
\usage{
gcs_cloud_auth(json_file = NULL, gcloud = FALSE, email = NULL)

gcs_get_cloud_auth()
}
\arguments{
\item{json_file}{character string. A JSON file that can be used to authenticate with
Google Cloud Storage. If the value is `NULL`, the current credential will be erased.}

\item{gcloud}{logical. Whether use gcloud to authenticate with Google Cloud Storage.
If the value is `TRUE`, the parameter `json_file` will be ignored.}

\item{email}{Character string or NULL. For gcloud only. Account to get the access token for. 
If not specified, the current active account in gcloud will be used.}
}
\value{
gcs_cloud_auth : No return value
gcs_get_cloud_auth : A list containing credentials information
}
\description{
Authenticate with Google Cloud Storage. You can download the JSON credential 
file from Google Gloud Platform. The credentials can be set globally by the 
evironment variable `GOOGLE_APPLICATION_CREDENTIALS` or `GCS_AUTH_FILE`.
}
\details{
When the package is loaded, it first searchs the credential file from the enviroment 
variable `GOOGLE_APPLICATION_CREDENTIALS`. If the credentials is not found, the environment variable
`GCS_AUTH_FILE` will be used intead. If both variables are not specified. Users need to specify the
credentials by calling `gcs_cloud_auth` function.
}
\examples{
## Authenticate with a service account
\dontrun{
gcs_cloud_auth(json_file = "/example.json")
}
}
