% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudFunctions.R
\name{gcs_set_read_buff}
\alias{gcs_set_read_buff}
\alias{gcs_set_write_buff}
\alias{gcs_get_read_buff}
\alias{gcs_get_write_buff}
\title{Get/Set read/write connection buffer size}
\usage{
gcs_set_read_buff(buff_size = 1024L * 1024L)

gcs_set_write_buff(buff_size = 1024L * 1024L)

gcs_get_read_buff()

gcs_get_write_buff()
}
\arguments{
\item{buff_size}{Integer. The buffer size}
}
\value{
Get functions: the current buffer size.
Set functions: the previous buffer size.
}
\description{
Get/Set read/write connection buffer size, the buffer size can be set at any time
but only takes effect on the connections created after the change. The default
value is 1024 *1024 bytes (1 Mega bytes) for both read and write connections.
}
\examples{
gcs_get_read_buff()
gcs_get_write_buff()
}
