% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\docType{class}
\name{FileClass-class}
\alias{FileClass-class}
\alias{.FileClass}
\title{File class}
\description{
The properties of file class object can be accessed via `$` and
`[[` operators.  The symbol `..` can be used to go to the parent
folder of a file class object.
}
