% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud-functions.R
\name{gcs_connection}
\alias{gcs_connection}
\title{Connection to google cloud storage}
\usage{
gcs_connection(
  description,
  open = "rb",
  encoding = getOption("encoding"),
  bucket = NULL,
  user_pay = gcs_get_user_pay()
)
}
\arguments{
\item{description}{character(1). The Google Cloud URL to the file 
that you want to connect to. If the value is a file path
(e.g. "folder1/folder2/myfile"), the bucket name will be provided
separately in the paramter `bucket`.}

\item{open}{character(1). A description of how to open the
connection.  See details for possible values. the default 
is "rb".}

\item{encoding}{character(1). The encoding of the input/output
stream of a connection.  Currently the parameter `encoding`
should be either `native.enc` or `UTF8`. see `?connections` for
more detail.}

\item{bucket}{character(1). The name of the bucket that the file is
located in. If not supplied, value in `gcs_get_global_bucket()`
will be used. If a full path to the file is provided in
`description`, this parameter will be ignored.}

\item{user_pay}{logical(1). Whether users should be responsible
for the cost associated with accessing the data. If the value
is `TRUE`, the charge will be sent to the billing project. See
`?gcs_get_billing_project()` for more details.}
}
\value{
A connection
}
\description{
This function creates an R connection to a file on google cloud
storage.  A service account credentials is required for accessing
private data, the credentials can be set via `gcs_cloud_auth`. If
the bucket requires Requester Pays, a billing project needs to be
set in `gcs_set_billing_project`.
}
\details{
Possible values for the argument `open` are the
    combination of the following characters:

"r" or "w" : read or write mode. The GCS connection cannot be in
 both read and write modes.

"t" or "b" : text or binary mode. If not specified, the default is
 text mode.
}
\examples{

## Open for reading the public Landsat data # on google cloud
## storage in text mode

f <- "gs://genomics-public-data/NA12878.chr20.sample.DeepVariant-0.7.2.vcf"
con <- gcs_connection(description = f, open = "rt")
readLines(con, n = 4L)
close(con)

}
