% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudFunctions.R
\name{gcs_dir}
\alias{gcs_dir}
\title{List bucket/folder/object}
\usage{
gcs_dir(path, delimiter = TRUE, recursive = FALSE, depth = 2L)
}
\arguments{
\item{path}{Character(1), the path to the bucket/folder/file,
seperated by `delimiter`.}

\item{delimiter}{Logical(1), whether to use `/` as a path delimiter}

\item{recursive}{Logical(1), whether recursively query all subdirectory.
If TRUE, all information of the subdirectories will be downloaded. The 
time cost can be significantly reduced if the value is FALSE. The parameter
only works with bucket/folder.}

\item{depth}{Integer(1), the depth of the reursive search}
}
\value{
A `FolderClass` object or a `FileClass` object
}
\description{
Get a list of objects in a bucket/folder if the path ends with `delimiter`.
Otherwise, get a description of a file.
}
\examples{
gcs_dir(path = "genomics-public-data/")
}
