% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud-functions.R
\name{gcs_set_billing_project}
\alias{gcs_set_billing_project}
\alias{gcs_get_billing_project}
\alias{gcs_get_requester_pays}
\alias{gcs_set_requester_pays}
\alias{gcs_is_requester_pays}
\title{Requester Pays}
\usage{
gcs_set_billing_project(billing_project = NULL, gcloud = FALSE)

gcs_get_billing_project()

gcs_get_requester_pays()

gcs_set_requester_pays(x)

gcs_is_requester_pays(bucket)
}
\arguments{
\item{billing_project}{character(1) or NULL. The project's ID which 
the charges will be sent to. If the value is not NULL, it will 
overwrite the default setting. See details.}

\item{gcloud}{logical. Whether use gcloud to authenticate with
Google Cloud Storage.  If the value is `TRUE`, the parameter
`json_file` will be ignored. See details.}

\item{x}{logical(1), whether the user should pay for the cost.}

\item{bucket}{character(1), the bucket name or uri}
}
\value{
gcs_get_billing_project: character(1) or NULL
gcs_get_requester_pays: logical(1)
}
\description{
These functions allow you to set billing project, change the default 
billing target and check if a bucket has Requester Pays enabled. See
the details section in `?authentication` for more information.
}
\examples{
gcs_get_billing_project()
gcs_get_requester_pays()
}
