% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\name{show,FileClass-method}
\alias{show,FileClass-method}
\title{Print object of class `FileClass`}
\usage{
\S4method{show}{FileClass}(object)
}
\arguments{
\item{object}{an object of class `FileClass`}
}
\value{
Invisible `Object`
}
\description{
Print object of class `FileClass`
}
