% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderClass.R
\name{show,FolderClass-method}
\alias{show,FolderClass-method}
\title{Print object of class `FolderClass`}
\usage{
\S4method{show}{FolderClass}(object)
}
\arguments{
\item{object}{an object of class `FolderClass`}
}
\value{
invisible NULL
}
\description{
Print object of class `FolderClass`
}
