% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileClass.R
\name{$,FileClass-method}
\alias{$,FileClass-method}
\alias{[[,FileClass-method}
\title{Get an element from `FileClass` object}
\usage{
\S4method{$}{FileClass}(x, name)

\S4method{[[}{FileClass}(x, i, exact = TRUE)
}
\arguments{
\item{x}{an object of class `FileClass`}

\item{name, i}{Character(1), the name of the element}

\item{exact}{Logical(1), Controls possible partial matching of `[[`
when extracting by a character(1)}
}
\value{
A `FolderClass` object or a `FileClass` object
}
\description{
Get an element from `FileClass` object
}
