% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FolderClass.R
\name{$,FolderClass-method}
\alias{$,FolderClass-method}
\alias{[[,FolderClass-method}
\title{Get an element from `FolderClass` object}
\usage{
\S4method{$}{FolderClass}(x, name)

\S4method{[[}{FolderClass}(x, i, exact = TRUE)
}
\arguments{
\item{x}{an object of class `FolderClass`}

\item{name, i}{Character(1), the name of the element}

\item{exact}{Logical(1), Controls possible partial matching
of `[[` when extracting by a character(1)}
}
\value{
A `FolderClass` object or a `FileClass` object
}
\description{
Get an element from `FolderClass` object
}
