% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcs_cloud_auth}
\alias{gcs_cloud_auth}
\title{Get/Set google credential file}
\usage{
gcs_cloud_auth(json_file)
}
\arguments{
\item{creds}{character string. A JSON file that can be used to authenticate
Google Cloud Storage.}
}
\description{
Authenticate with Google Cloud Storage. You can download the JSON credential 
file from Google Gloud Platform. The credentials can be set globally by the 
evironment variable `GOOGLE_APPLICATION_CREDENTIALS` or `GCS_AUTH_FILE`.
}
\details{
When the package is loaded, it first searchs the credential file from the enviroment 
variable `GOOGLE_APPLICATION_CREDENTIALS`. If the credentials is not found, the environment variable
`GCS_AUTH_FILE` will be used intead. If both variables are not specified. Users need to specify the
credentials by calling `gcs_cloud_auth` function.
}
\examples{
## Get the default credential file
gcs_get_cloud_auth()
}
