% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gcs_input_stream_buff}
\alias{gcs_input_stream_buff}
\alias{gcs_output_stream_buff}
\alias{gcs_get_input_stream_buff}
\alias{gcs_get_output_stream_buff}
\title{Get/Set IO stream buffer size}
\usage{
gcs_input_stream_buff(buff_size)

gcs_output_stream_buff(buff_size)

gcs_get_input_stream_buff()

gcs_get_output_stream_buff()
}
\arguments{
\item{buff_size}{Integer. The buffer size}
}
\description{
Get/Set IO stream buffer size, the buffer size can be set at any time
but only takes effect on the connections created after the change. The default 
value is 1024 *1024 bytes (1 Mega bytes) for both read and write connections.
}
\examples{
gcs_get_input_stream_buff()
gcs_get_output_stream_buff()
}
