% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_classes.R
\docType{class}
\name{FolderClass-class}
\alias{FolderClass-class}
\alias{.FolderClass}
\title{Folder class}
\description{
View and access files. You can change the current directory by `$`
and `[[` operators. The symbol `..` can be used to go to the parent
folder of a folder object.
}
