% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefinitions.R
\docType{class}
\name{FolderClass-class}
\alias{FolderClass-class}
\alias{.FolderClass}
\title{Folder class}
\description{
view and access files. The Folder class object can be accessed
 via `$` and `[[` operators.
}
