% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudConnection.R
\name{gcs_connection}
\alias{gcs_connection}
\title{Connection to google cloud storage}
\usage{
gcs_connection(description, open, encoding = getOption("encoding"),
  credentials = NULL, bucket = NULL)
}
\arguments{
\item{description}{character string. The name of the file that you want to connect to.
It can be either the file name or a full path to the file.}

\item{open}{character string. A description of how to open the connection.
See `?connections` for possible values.}

\item{encoding}{character string. The character encoding. Currently only support either `Unicode`
or `UTF8`}

\item{credentials}{character string. A file path to the JSON credentials. If not supplied, the default
value in `gcs_get_cloud_auth()` will be used. If the credentials is an empty string(AKA ""), an anonymous 
credentials will be supplied.}

\item{bucket}{character string. The name of the bucket that the file is located in. If not supplied, 
value in `gcs_get_global_bucket()` will be used. If a full path to the file is provided in `description`,
this parameter will be ignored.}
}
\description{
This function creates an R connection to a file on google cloud storage.
}
\examples{
## Connect to the Landsat data on google cloud storage
## This is a public dataset so there is no need to provide
## credentials.
file <- "gs://genomics-public-data/NA12878.chr20.sample.DeepVariant-0.7.2.vcf"
con <- gcs_connection(description = file, open = "r", credentials = "")
readLines(con, n = 1L)
close(con)

}
