% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudFunction.R
\name{gcs_cp}
\alias{gcs_cp}
\title{copy files to and from buckets}
\usage{
gcs_cp(from, to)
}
\arguments{
\item{from, to}{Character. The path of the file. It can be either a file path
or a google URI.}
}
\value{
No return value
}
\description{
The function supports moving files from bucket to bucket/ disk to bucket/ 
bucket to disk. Note that the existing destination file will be overwritten.
}
\examples{
# Download a file to a disk
#gcs_cp("gs://bucket_name/file_name","file_path_on_disk")
}
