% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud-functions.R
\name{gcs_cp}
\alias{gcs_cp}
\title{copy files to and from buckets}
\usage{
gcs_cp(from, to, recursive = TRUE, user_pay = gcs_get_user_pay())
}
\arguments{
\item{from, to}{Character(1). The path to the folder/file.  At least
one path must be a google URL. It is recommended to explicitly
add a trailing "/" if the parameter is a path to a folder.}

\item{recursive}{logical(1). Whether to recursively copy the files in
the subfolders.}

\item{user_pay}{logical(1). Whether users should be responsible
for the cost associated with accessing the data. If the value
is `TRUE`, the charge will be sent to the billing project. See
`?gcs_get_billing_project()` for more details.}
}
\value{
No return value
}
\description{
The function supports moving files or folders from bucket to
bucket, disk to bucket and bucket to disk.  Note that the existing
destination file will be overwritten.
}
\examples{
tmp_path <- tempdir()

## Download a file to a disk
gcs_cp("gs://genomics-public-data/NA12878.chr20.sample.bam", tmp_path)

## Check the file existance
file.exists(file.path(tmp_path, "NA12878.chr20.sample.bam"))

## Download all files in a path.
## The files in the subfolders will not be copied due to `recursive = FALSE`
folder_path <- file.path(tmp_path, "example")
gcs_cp("gs://genomics-public-data/", folder_path, recursive = FALSE)

## Check the file existance
list.files(folder_path)

}
