% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudFunction.R
\name{gcs_dir}
\alias{gcs_dir}
\title{List bucket/object}
\usage{
gcs_dir(bucket, file = NULL, max_files = NULL)
}
\arguments{
\item{bucket}{Character(1), the name of the bucket}

\item{file}{character(1) or NULL, the name of the file}

\item{max_files}{integer or NULL, The maximum number of objects to 
return in a List request.}
}
\description{
Get a list of objects in a bucket, or get a description of a file.
}
